+++ src/app/workbench/window-manager/window-frame.component.ts
@@ -6,8 +6,12 @@ import {
   OnInit,
   Output,
 } from '@angular/core';
+import { NgStyle } from '@angular/common';
+import { IonicModule } from '@ionic/angular';
 import { HoroRequest, ProcessRequest } from 'src/app/type/interface/request-data';
 import { Mode } from 'src/app/native/enum';
+import { NativePageModule } from 'src/app/native/native.module';
+import { ProcessPageModule } from 'src/app/process/process.module';
 import { ProcessName } from 'src/app/process/enum/process';
 import { ProfectionMode } from 'src/app/process/enum/profection-mode';
 import {
@@ -21,7 +25,13 @@ import {
   selector: 'app-window-frame',
   templateUrl: './window-frame.component.html',
   styleUrls: ['./window-frame.component.scss'],
-  standalone: false,
+  standalone: true,
+  imports: [
+    NgStyle,
+    IonicModule,
+    NativePageModule,
+    ProcessPageModule,
+  ],
 })
 export class WindowFrameComponent implements OnInit, OnDestroy {
   @Input() window!: WorkbenchWindow;